test(react): clean up persona-coin test and use isConformant without side-effects
