+++ change/@fluentui-react-da22778f-cbbf-4200-b03c-752627517a9c.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "test(react): clean up persona-coin test and use isConformant without side-effects",
+  "packageName": "@fluentui/react",
+  "email": "martinhochel@microsoft.com",
+  "dependentChangeType": "none"
+}
